+++ vanadium-benches/Cargo.toml
@@ -7,4 +7,4 @@ edition = "2018"
 
 [dependencies.vanadium-core]
 path = "../vanadium-core"
-features = ["glommio-support", "tokio-uring-support", "tokio-support"]
+features = ["glommio-support"]
