+++ vanadium-benches/src/bin/big-glommio.rs
@@ -1,8 +1,7 @@
 use vanadium_benches::{bench_image, large_header};
-use vanadium_core::asynchronous_ops::glommio::bip::GlommioBip;
+use vanadium_core::image_backends::get_image_f32;
 
 fn main() {
-    let mut image: GlommioBip<f32> = GlommioBip::new(large_header());
-
-    bench_image(&mut image)
+    let mut image = get_image_f32(Some("glommio"), large_header()).unwrap();
+    bench_image(image.as_mut())
 }
