+++ vanadium-benches/src/bin/small-glommio.rs
@@ -1,8 +1,7 @@
 use vanadium_benches::{bench_image, small_header};
-use vanadium_core::asynchronous_ops::glommio::bip::GlommioBip;
+use vanadium_core::image_backends::get_image_f32;
 
 fn main() {
-    let mut image: GlommioBip<f32> = GlommioBip::new(small_header());
-
-    bench_image(&mut image)
+    let mut image = get_image_f32(Some("glommio"), small_header()).unwrap();
+    bench_image(image.as_mut())
 }
