+++ vanadium-benches/src/bin/small-sync.rs
@@ -1,8 +1,7 @@
 use vanadium_benches::{bench_image, small_header};
-use vanadium_core::sync_syscall::SyncBip;
+use vanadium_core::image_backends::get_image_f32;
 
 fn main() {
-    let mut image: SyncBip<f32> = SyncBip::new(small_header()).unwrap();
-
-    bench_image(&mut image)
+    let mut image = get_image_f32(Some("sync"), small_header()).unwrap();
+    bench_image(image.as_mut())
 }
