+++ vanadium-core/Cargo.toml
@@ -3,36 +3,47 @@ name = "vanadium-core"
 version = "0.1.0"
 edition = "2018"
 
+###################################################################################################
+# Features
+###################################################################################################
+
 [features]
-async = ["futures"]
 header-parsing = ["serde"]
-tokio-support = ["tokio", "async"]
-tokio-uring-support = ["tokio-uring", "async"]
-glommio-support = ["glommio"]
+tokio-uring-support = ["tokio-uring", "futures"]
+glommio-support = ["glommio", "futures"]
+
+###################################################################################################
+# Dependencies
+###################################################################################################
 
 [dependencies]
 num-traits = "^0.2.14"
 nalgebra = "^0.28.0"
 byteorder = "^1.4"
+cfg-if = "^1.0.0"
 
-[dependencies.futures]
-version = "^0.3.15"
-optional = true
-
+###################################################################################################
+# Not always needed for libraries
+###################################################################################################
 [dependencies.serde]
 version = "^1.0.0"
 features = ["derive"]
 optional = true
 
-[dependencies.tokio]
-version = "^1.9.0"
-features = ["sync"]
+###################################################################################################
+# Async backends
+###################################################################################################
+[dependencies.futures]
+version = "^0.3.15"
 optional = true
 
 [dependencies.tokio-uring]
-version = "0.1.0"
+version = "^0.1.0"
 optional = true
 
+###################################################################################################
+# Unix only
+###################################################################################################
 [dependencies.glommio]
 version = "^0.5.0"
 optional = true
