+++ vanadium-core/src/backends/mod.rs
@@ -2,11 +2,14 @@ use std::error::Error;
 
 use nalgebra::{ComplexField, DMatrix, Dynamic, SymmetricEigen};
 
-#[cfg(feature = "tokio-support")]
-pub mod tokio;
+use sync_syscall::bip::SyncBip;
+use sync_syscall::bsq::SyncBsq;
 
-#[cfg(feature = "tokio-uring-support")]
-pub mod tokio_uring;
+use crate::backends::glommio::bip::GlommioBip;
+use crate::headers::{Header, ImageFormat};
+
+// #[cfg(feature = "tokio-uring-support")]
+// pub mod tokio_uring;
 
 #[cfg(feature = "glommio-support")]
 pub mod glommio;
@@ -14,6 +17,21 @@ pub mod glommio;
 pub mod sync_syscall;
 
 pub type GenericResult<T> = Result<T, Box<dyn Error>>;
+
+pub fn get_image_f32(backend: Option<&str>, header: Header)
+                     -> Result<Box<dyn Image<f32>>, Box<dyn Error>>
+{
+    Ok(if let Some(s) = backend {
+        match (s, header.format) {
+            ("glommio", ImageFormat::Bip) => Box::new(GlommioBip::new(header)),
+            ("sync", ImageFormat::Bip) => Box::new(SyncBip::new(header)?),
+            ("sync", ImageFormat::Bsq) => Box::new(SyncBsq::new(header)?),
+            _ => unimplemented!()
+        }
+    } else {
+        todo!()
+    })
+}
 
 pub trait Image<T> where T: ComplexField {
     fn means(&mut self) -> GenericResult<Vec<T>>;
