+++ vanadium-core/src/backends/sync_syscall/bsq.rs
@@ -5,24 +5,18 @@ use std::{mem, io};
 use byteorder::{LittleEndian, ReadBytesExt};
 use nalgebra::{DMatrix, Dynamic, SymmetricEigen};
 
-use crate::backends::{GenericResult, Image};
-use crate::specialization::bip::Bip;
+use crate::{GenericResult, Image};
 use crate::specialization::bsq::Bsq;
 use crate::headers::Header;
 
-pub struct SyncBip<T> {
-    pub (crate) file: File,
-    pub (crate) bip: Bip<T>,
-}
-
 pub struct SyncBsq<T> {
-    pub (crate) file: File,
-    pub (crate) bsq: Bsq<T>,
+    file: File,
+    bsq: Bsq<T>,
 }
 
-impl <T> SyncBip<T> {
+impl <T> SyncBsq<T> {
     pub fn new(header: Header) -> io::Result<Self> {
-        let bip = Bip {
+        let bsq = Bsq {
             dims: header.dims,
             phantom: Default::default()
         };
@@ -31,61 +25,8 @@ impl <T> SyncBip<T> {
 
         Ok(Self {
             file,
-            bip,
+            bsq,
         })
-    }
-}
-
-impl Image<f32> for SyncBip<f32> {
-    fn means(&mut self) -> GenericResult<Vec<f32>> {
-        self.file.seek(SeekFrom::Start(0))?;
-        let mut accumulator = vec![0.0; self.bip.pixel_length()];
-        let mut buffer = vec![0.0; self.bip.pixel_length()];
-
-        for _ in 0..self.bip.num_pixels() {
-            self.file.read_f32_into::<LittleEndian>(&mut buffer)?;
-            self.bip.map_mean(&buffer, &mut accumulator);
-        }
-
-        self.bip.reduce_mean(&mut accumulator);
-
-        Ok(accumulator)
-    }
-
-    fn std_deviations(&mut self, means: &[f32]) -> GenericResult<Vec<f32>> {
-        self.file.seek(SeekFrom::Start(0))?;
-
-        let mut accumulator = vec![0.0; self.bip.pixel_length()];
-        let mut buffer = vec![0.0; self.bip.pixel_length()];
-
-        for _ in 0..self.bip.num_pixels() {
-            self.file.read_f32_into::<LittleEndian>(&mut buffer)?;
-            self.bip.map_std_dev(&buffer, means, &mut accumulator);
-        }
-
-        self.bip.reduce_std_dev(&mut accumulator);
-
-        Ok(accumulator)
-    }
-
-    fn covariance_matrix(&mut self, means: &[f32], std_devs: &[f32]) -> GenericResult<DMatrix<f32>> {
-        self.file.seek(SeekFrom::Start(0))?;
-
-        let mut accumulator = DMatrix::zeros(self.bip.dims.channels, self.bip.dims.channels);
-        let mut buffer = vec![0.0; self.bip.pixel_length()];
-
-        for _ in 0..self.bip.num_pixels() {
-            self.file.read_f32_into::<LittleEndian>(&mut buffer)?;
-            self.bip.map_covariance(&mut buffer, means, std_devs, &mut accumulator);
-        }
-
-        self.bip.reduce_covariance(&mut accumulator);
-
-        Ok(accumulator)
-    }
-
-    fn write_standardized(&mut self, _path: &str, _means: &[f32], _std_devs: &[f32], _eigen: &SymmetricEigen<f32, Dynamic>) -> GenericResult<()> {
-        todo!()
     }
 }
