+++ vanadium-core/src/backends/sync_syscall/mod.rs
@@ -1,0 +1,2 @@
+pub mod bip;
+pub mod bsq;
