+++ vanadium-core/src/lib.rs
@@ -12,14 +12,6 @@ mod backends;
 
 mod util;
 
-#[cfg(feature = "async")]
-pub mod asynchronous_ops {
-    #[cfg(feature = "glommio-support")]
-    pub use crate::backends::glommio;
-    #[cfg(feature = "tokio-uring-support")]
-    pub use crate::backends::tokio_uring;
-}
-
-pub mod sync_syscall {
-    pub use crate::backends::sync_syscall::*;
+pub mod image_backends {
+    pub use crate::backends::get_image_f32;
 }
